Split caching-related business logic into parameter_cache.go
